redone build process
Completely overhauled the build / release process for version updates. Now there are more properties to add in local.properties but you no longer have to use the web UI at all. Also we are no longer relying on an outdated plugin/script that someone wrote and has not updated. we are using the official bintray plugin for gradle